Move the various project-affiliated code files to their own module
