+++ src/project/actions/buffer_clone_shallow.rs
@@ -9,7 +9,7 @@ use redo::Command;
 use serde::{Serialize, Deserialize};
 use simple_error::{SimpleResult, SimpleError, bail};
 
-use crate::h2project::H2Project;
+use crate::project::h2project::H2Project;
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct ActionBufferCloneShallowForward {
@@ -114,8 +114,8 @@ mod tests {
 
     use crate::transformation::TransformHex;
 
-    use crate::action::Action;
-    use crate::h2project::H2Project;
+    use crate::project::actions::Action;
+    use crate::project::h2project::H2Project;
 
     #[test]
     fn test_action() -> SimpleResult<()> {
