+++ src/project/actions/buffer_create_empty.rs
@@ -1,46 +1,45 @@
-//! Create a new buffer consisting of the provided data. Will commonly be read
-//! from a file.
+//! Create a new buffer of a particular size, initialized to all zeroes.
 
 use redo::Command;
 use serde::{Serialize, Deserialize};
 use simple_error::{SimpleResult, SimpleError, bail};
 
-use crate::h2project::H2Project;
-use crate::h2buffer::H2Buffer;
+use crate::project::h2project::H2Project;
+use crate::project::h2buffer::H2Buffer;
 
 #[derive(Serialize, Deserialize, Debug)]
-pub struct ActionBufferCreateFromBytesForward {
+pub struct ActionBufferCreateEmptyForward {
     pub name: String,
-    pub data: Vec<u8>,
+    pub size: usize,
     pub base_address: usize,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
-struct ActionBufferCreateFromBytesBackward {
+struct ActionBufferCreateEmptyBackward {
     name: String,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
-pub struct ActionBufferCreateFromBytes {
-    forward: Option<ActionBufferCreateFromBytesForward>,
-    backward: Option<ActionBufferCreateFromBytesBackward>,
+pub struct ActionBufferCreateEmpty {
+    forward: Option<ActionBufferCreateEmptyForward>,
+    backward: Option<ActionBufferCreateEmptyBackward>,
 }
 
-impl ActionBufferCreateFromBytes {
-    pub fn new(forward: ActionBufferCreateFromBytesForward) -> Self {
-        ActionBufferCreateFromBytes {
+impl ActionBufferCreateEmpty {
+    pub fn new(forward: ActionBufferCreateEmptyForward) -> Self {
+        ActionBufferCreateEmpty {
             forward: Some(forward),
             backward: None,
         }
     }
 }
 
-impl From<(&str, Vec<u8>, usize)> for ActionBufferCreateFromBytes {
-    fn from(o: (&str, Vec<u8>, usize)) -> Self {
-        ActionBufferCreateFromBytes {
-            forward: Some(ActionBufferCreateFromBytesForward {
+impl From<(&str, usize, usize)> for ActionBufferCreateEmpty {
+    fn from(o: (&str, usize, usize)) -> Self {
+        ActionBufferCreateEmpty {
+            forward: Some(ActionBufferCreateEmptyForward {
                 name: o.0.to_string(),
-                data: o.1,
+                size: o.1,
                 base_address: o.2,
             }),
             backward: None,
@@ -48,7 +47,7 @@ impl From<(&str, Vec<u8>, usize)> for ActionBufferCreateFromBytes {
     }
 }
 
-impl Command for ActionBufferCreateFromBytes {
+impl Command for ActionBufferCreateEmpty {
     type Target = H2Project;
     type Error = SimpleError;
 
@@ -59,14 +58,13 @@ impl Command for ActionBufferCreateFromBytes {
             None => bail!("Failed to apply: missing context"),
         };
 
-        // Sanity check: it has a size
         // Apply the change
-        let buffer = H2Buffer::new(forward.data.clone(), forward.base_address)?;
+        let buffer = H2Buffer::new(vec![0; forward.size], forward.base_address)?;
         project.buffer_insert(&forward.name, buffer)?;
 
-        // Populate backward for undo
-        self.backward = Some(ActionBufferCreateFromBytesBackward {
-            name: forward.name.clone(),
+        // Swap backward + forward
+        self.backward = Some(ActionBufferCreateEmptyBackward {
+            name: forward.name.to_string(),
         });
         self.forward = None;
 
@@ -82,9 +80,9 @@ impl Command for ActionBufferCreateFromBytes {
         let name = &backward.name;
         let buffer = project.buffer_remove(name)?;
 
-        self.forward = Some(ActionBufferCreateFromBytesForward {
+        self.forward = Some(ActionBufferCreateEmptyForward {
             name: name.clone(),
-            data: buffer.data,
+            size: buffer.data.len(),
             base_address: buffer.base_address,
         });
         self.backward = None;
@@ -98,27 +96,28 @@ mod tests {
     use super::*;
     use simple_error::SimpleResult;
 
-    use crate::h2project::H2Project;
+    use crate::project::h2project::H2Project;
     use redo::Record;
     use pretty_assertions::assert_eq;
 
     #[test]
     fn test_action() -> SimpleResult<()> {
-        let mut record: Record<ActionBufferCreateFromBytes> = Record::new(
+        let mut record: Record<ActionBufferCreateEmpty> = Record::new(
             H2Project::new("name", "1.0")
         );
 
         assert_eq!(0, record.target().buffers().len());
 
-        record.apply(("buffer", vec![0, 1, 2, 4], 0x80000000).into())?;
+        record.apply(("buffer", 10, 0x80000000).into())?;
 
         let buffers = record.target().buffers();
         assert_eq!(1, buffers.len());
-        assert_eq!(4, buffers["buffer"].data.len());
-        assert_eq!(vec![0, 1, 2, 4], buffers["buffer"].data);
+
+        assert_eq!(10, buffers["buffer"].data.len());
         assert_eq!(0x80000000, buffers["buffer"].base_address);
 
         record.undo()?;
+
         record.redo()?;
 
         Ok(())
@@ -126,25 +125,25 @@ mod tests {
 
     #[test]
     fn test_action_fails_if_buffer_already_exists() -> SimpleResult<()> {
-        let mut record: Record<ActionBufferCreateFromBytes> = Record::new(
+        let mut record: Record<ActionBufferCreateEmpty> = Record::new(
             H2Project::new("name", "1.0")
         );
 
-        assert!(record.apply(("buffer", vec![0, 1, 2, 4], 0x80000000).into()).is_ok());
-        assert!(record.apply(("buffer123", vec![0, 1, 2, 4], 0x80000000).into()).is_ok());
-        assert!(record.apply(("buffer123", vec![0, 1, 2, 4], 0x80000000).into()).is_err());
-        assert!(record.apply(("buffer", vec![0, 1, 2, 4], 0x80000000).into()).is_err());
+        assert!(record.apply(("buffer", 10, 0x80000000).into()).is_ok());
+        assert!(record.apply(("buffer2", 10, 0x80000000).into()).is_ok());
+        assert!(record.apply(("buffer", 10, 0x80000000).into()).is_err());
+        assert!(record.apply(("buffer2", 10, 0x80000000).into()).is_err());
 
         Ok(())
     }
 
     #[test]
     fn test_action_fails_with_zero_size() -> SimpleResult<()> {
-        let mut record: Record<ActionBufferCreateFromBytes> = Record::new(
+        let mut record: Record<ActionBufferCreateEmpty> = Record::new(
             H2Project::new("name", "1.0")
         );
 
-        assert!(record.apply(("buffer", vec![], 0x80000000).into()).is_err());
+        assert!(record.apply(("buffer", 0, 0x80000000).into()).is_err());
 
         Ok(())
     }
