+++ src/project/actions/buffer_transform.rs
@@ -13,7 +13,7 @@ use simple_error::{SimpleResult, SimpleError, bail};
 
 use crate::transformation::Transformation;
 
-use crate::h2project::H2Project;
+use crate::project::h2project::H2Project;
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct ActionBufferTransformForward {
@@ -104,10 +104,10 @@ impl Command for ActionBufferTransform {
 mod tests {
     use simple_error::SimpleResult;
 
-    use crate::h2project::H2Project;
+    use crate::project::h2project::H2Project;
     use redo::Record;
     use pretty_assertions::assert_eq;
-    use crate::action::Action;
+    use crate::project::actions::Action;
     use crate::transformation::{TransformHex, TransformBase64};
 
     #[test]
