+++ src/project/actions/mod.rs
@@ -9,7 +9,7 @@ use simple_error::{SimpleResult, SimpleError};
 use std::ops::Range;
 
 use crate::transformation::Transformation;
-use crate::h2project::H2Project;
+use crate::project::h2project::H2Project;
 
 // Still needed:
 // * Layers
